Extract bounded-stderr diagnostics into a shared mixin
codex and pi duplicated the bounded stderr buffer near-verbatim: the buffer
itself, appendBounded with line+byte caps, the onFinalize drain-join, and
diagnosticContext. Hoist it into BufferedStderrDiagnostics in tools; each
backend keeps its own handleStderr (the noise filters genuinely differ) and
calls recordStderr. pi's onFinalize chains super.onFinalize() then closes its
per-turn temp resources. AppendBoundedTest moves from codex to tools.
